requirements.txt
streamlit
pillow
piexif
pandas